Feat insert method
This method inserts a new node into the binary search tree, with the gi-
ven value. If that value is already inside the tree, it won't add anyth-
ing.

It adds it to the appropriate area by checking if smaller it'll go left,
and if bigger it'll go right, until an empty spot is available.