Feat preorder trversal
With the method #preorder travsersal, it is possible to traverse the tr-
ee in preorder manner. This means we look at the node itself, then the
one to its left, then the one to its right.

If a block is given, it'll yield in preorder sequence. Otherwise, it'll
just return an array including the values of the tree in breorder seque-
nce.