Fix issue with preorder when blcok given
When a block was given to this method, it would only yield the first va-
lue. Now it it creates an array of the tree's nodes in preorder sequenc-
e. If a block was given, it'll iterate over the array and yield each it-
em. Else, it'll just return the array.